# openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/api/phones/":{"post":{"tags":["phones"],"summary":"Create","operationId":"create_api_phones__post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PhoneCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Phone"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["phones"],"summary":"Read All","operationId":"read_all_api_phones__get","parameters":[{"name":"skip","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Phone"},"title":"Response Read All Api Phones  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/phones/{phone_id}":{"get":{"tags":["phones"],"summary":"Read","operationId":"read_api_phones__phone_id__get","parameters":[{"name":"phone_id","in":"path","required":true,"schema":{"type":"integer","title":"Phone Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Phone"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["phones"],"summary":"Delete","operationId":"delete_api_phones__phone_id__delete","parameters":[{"name":"phone_id","in":"path","required":true,"schema":{"type":"integer","title":"Phone Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/incidents/":{"post":{"tags":["incidents"],"summary":"Create","operationId":"create_api_incidents__post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/IncidentCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Incident"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["incidents"],"summary":"Read All","operationId":"read_all_api_incidents__get","parameters":[{"name":"skip","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Incident"},"title":"Response Read All Api Incidents  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/incidents/{incident_id}":{"get":{"tags":["incidents"],"summary":"Read","operationId":"read_api_incidents__incident_id__get","parameters":[{"name":"incident_id","in":"path","required":true,"schema":{"type":"integer","title":"Incident Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Incident"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["incidents"],"summary":"Delete","operationId":"delete_api_incidents__incident_id__delete","parameters":[{"name":"incident_id","in":"path","required":true,"schema":{"type":"integer","title":"Incident Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Incident":{"properties":{"group_id":{"type":"integer","title":"Group Id"},"trigger_phone_id":{"type":"integer","title":"Trigger Phone Id"},"photo":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Photo"},"id_incident":{"type":"integer","title":"Id Incident"},"time":{"type":"string","format":"date-time","title":"Time"}},"type":"object","required":["group_id","trigger_phone_id","photo","id_incident","time"],"title":"Incident"},"IncidentCreate":{"properties":{"group_id":{"type":"integer","title":"Group Id"},"trigger_phone_id":{"type":"integer","title":"Trigger Phone Id"},"photo":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Photo"}},"type":"object","required":["group_id","trigger_phone_id","photo"],"title":"IncidentCreate"},"Phone":{"properties":{"geolocation":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Geolocation"},"status":{"type":"boolean","title":"Status","default":true},"timestatus":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Timestatus"},"group_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Group Id"},"id_phone":{"type":"integer","title":"Id Phone"}},"type":"object","required":["geolocation","group_id","id_phone"],"title":"Phone"},"PhoneCreate":{"properties":{"geolocation":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Geolocation"},"status":{"type":"boolean","title":"Status","default":true},"timestatus":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Timestatus"},"group_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Group Id"}},"type":"object","required":["geolocation","group_id"],"title":"PhoneCreate"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}